Fix for panel behavior when browser window size is changed
