+++ src/content-handlers/iiif/BaseConfig.ts
@@ -169,68 +169,6 @@ type HeaderPanel = ModuleConfig & {
 
 type LeftPanel = ExpandPanel & {};
 
-type SearchLeftPanelOptions = DialogueOptions &
-  ExpandPanelOptions & {
-    /** Order in which canvases are displayed */
-    canvasDisplayOrder: string;
-    /** Canvases to exclude from display */
-    canvasExclude: string;
-    /** Determines if copying to clipboard is enabled */
-    copyToClipboardEnabled: boolean;
-    /** Determines if download is enabled */
-    limitToRange: boolean;
-    /** Order in which manifests are displayed */
-    manifestDisplayOrder: string;
-    /** Manifests to exclude from display */
-    manifestExclude: string;
-    /** Language codes for right-to-left languages */
-    rtlLanguageCodes: string;
-    /** Determines if all languages should be shown */
-    showAllLanguages: boolean;
-    /** Limit for the text */
-    textLimit: number;
-    /** Type of the text limit */
-    textLimitType: string;
-  };
-
-type SearchLeftPanelContent = DialogueContent &
-  ExpandPanelContent & {
-    attribution: string;
-    canvasHeader: string;
-    collapse: string;
-    collapseFull: string;
-    copiedToClipboard: string;
-    copyToClipboard: string;
-    description: string;
-    expand: string;
-    expandFull: string;
-    holdingText: string;
-    less: string;
-    license: string;
-    logo: string;
-    manifestHeader: string;
-    more: string;
-    noData: string;
-    page: string;
-    rangeHeader: string;
-    title: string;
-    enterKeyword: string;
-    searchWithin: string;
-    clearSearch: string;
-    doSearch: string;
-    instanceFound: string;
-    instancesFound: string;
-    noMatches: string;
-    nextResult: string;
-    previousResult: string;
-    searchHitOf: string;
-  };
-
-export type SearchLeftPanel = ModuleConfig & {
-  options: SearchLeftPanelOptions;
-  content: SearchLeftPanelContent;
-};
-
 export type CenterPanelOptions = {
   titleEnabled: boolean;
   subtitleEnabled: boolean;
@@ -530,61 +468,6 @@ export type MoreInfoRightPanel = ModuleConfig & {
   content: MoreInfoRightPanelContent;
 };
 
-type TextRightPanelOptions = DialogueOptions &
-  ExpandPanelOptions & {
-    /** Order in which canvases are displayed */
-    canvasDisplayOrder: string;
-    /** Canvases to exclude from display */
-    canvasExclude: string;
-    /** Determines if copying to clipboard is enabled */
-    copyToClipboardEnabled: boolean;
-    /** Determines if download is enabled */
-    limitToRange: boolean;
-    /** Order in which manifests are displayed */
-    manifestDisplayOrder: string;
-    /** Manifests to exclude from display */
-    manifestExclude: string;
-    /** Language codes for right-to-left languages */
-    rtlLanguageCodes: string;
-    /** Determines if all languages should be shown */
-    showAllLanguages: boolean;
-    /** Limit for the text */
-    textLimit: number;
-    /** Type of the text limit */
-    textLimitType: string;
-  };
-
-type TextRightPanelContent = DialogueContent &
-  ExpandPanelContent & {
-    attribution: string;
-    canvasHeader: string;
-    collapse: string;
-    collapseFull: string;
-    copiedToClipboard: string;
-    copyToClipboard: string;
-    description: string;
-    expand: string;
-    expandFull: string;
-    holdingText: string;
-    less: string;
-    license: string;
-    logo: string;
-    manifestHeader: string;
-    more: string;
-    noData: string;
-    page: string;
-    rangeHeader: string;
-    title: string;
-    leftPage: string;
-    rightPage: string;
-    textNotFound: string;
-  };
-
-export type TextRightPanel = ModuleConfig & {
-  options: TextRightPanelOptions;
-  content: TextRightPanelContent;
-};
-
 export type Content = {
   authCORSError: string;
   authorisationFailedMessage: string;
@@ -611,11 +494,9 @@ export type BaseConfig = {
     headerPanel: HeaderPanel;
     helpDialogue: HelpDialogue;
     leftPanel: LeftPanel;
-    searchLeftPanel: SearchLeftPanel;
     loginDialogue: LoginDialogue;
     mobileFooterPanel: FooterPanel;
     moreInfoRightPanel: MoreInfoRightPanel;
-    textRightPanel: TextRightPanel;
     restrictedDialogue: RestrictedDialogue;
     settingsDialogue: SettingsDialogue;
     shareDialogue: ShareDialogue;
