+++ src/content-handlers/iiif/extensions/uv-av-extension/config/config.json
@@ -8,7 +8,6 @@
     "footerPanelEnabled": true,
     "headerPanelEnabled": true,
     "leftPanelEnabled": true,
-    "searchLeftPanelEnabled": true,
     "limitLocales": false,
     "metrics": [
       {
@@ -37,7 +36,6 @@
     "pessimisticAccessControl": false,
     "preserveViewport": false,
     "rightPanelEnabled": true,
-    "textRightPanelEnabled": true,
     "saveUserSettings": false,
     "clickToZoomEnabled": true,
     "searchWithinEnabled": true,
@@ -62,58 +60,6 @@
         "collapseFull": "$collapseFull",
         "expand": "$expandContents",
         "expandFull": "$expandGallery"
-      }
-    },
-    "searchLeftPanel": {
-      "options": {
-        "canvasDisplayOrder": "",
-        "canvasExclude": "",
-        "copyToClipboardEnabled": false,
-        "expandFullEnabled": false,
-        "limitToRange": false,
-        "manifestDisplayOrder": "",
-        "manifestExclude": "",
-        "panelAnimationDuration": 250,
-        "panelCollapsedWidth": 30,
-        "panelExpandedWidth": 255,
-        "panelOpen": false,
-        "rtlLanguageCodes": "ar, ara, dv, div, he, heb, ur, urd",
-        "showAllLanguages": false,
-        "textLimit": 4,
-        "textLimitType": "lines",
-        "topCloseButtonEnabled": false
-      },
-      "content": {
-        "attribution": "$attribution",
-        "canvasHeader": "$aboutTheImage",
-        "clearSearch": "$clearSearch",
-        "doSearch": "$doSearch",
-        "close": "$close",
-        "collapse": "$collapseSearch",
-        "collapseFull": "$collapseFull",
-        "copiedToClipboard": "$copiedToClipboard",
-        "copyToClipboard": "$copyToClipboard",
-        "description": "$description",
-        "enterKeyword": "$enterKeyword",
-        "expand": "$expandSearch",
-        "expandFull": "$expandGallery",
-        "holdingText": "$moduleGoesHere",
-        "less": "$less",
-        "license": "$license",
-        "logo": "$logo",
-        "manifestHeader": "$aboutTheItem",
-        "more": "$more",
-        "noData": "$noDataToDisplay",
-        "page": "$page",
-        "rangeHeader": "$aboutThisSection",
-        "searchWithin": "$searchWithinItem",
-        "title": "$search",
-        "instanceFound": "$instanceFound",
-        "instancesFound": "$instancesFound",
-        "noMatches": "$noMatches",
-        "nextResult": "$nextResult",
-        "previousResult": "$previousResult",
-        "searchHitOf": "$searchHitOf"
       }
     },
     "contentLeftPanel": {
@@ -312,51 +258,6 @@
         "page": "$page",
         "rangeHeader": "$aboutThisSection",
         "title": "$moreInformation"
-      }
-    },
-    "textRightPanel": {
-      "options": {
-        "canvasDisplayOrder": "",
-        "canvasExclude": "",
-        "copyToClipboardEnabled": false,
-        "expandFullEnabled": false,
-        "limitToRange": false,
-        "manifestDisplayOrder": "",
-        "manifestExclude": "",
-        "panelAnimationDuration": 250,
-        "panelCollapsedWidth": 30,
-        "panelExpandedWidth": 442,
-        "panelOpen": false,
-        "rtlLanguageCodes": "ar, ara, dv, div, he, heb, ur, urd",
-        "showAllLanguages": false,
-        "textLimit": 4,
-        "textLimitType": "lines",
-        "topCloseButtonEnabled": false
-      },
-      "content": {
-        "attribution": "$attribution",
-        "canvasHeader": "$aboutTheImage",
-        "close": "$close",
-        "collapse": "$collapseTranscribedText",
-        "collapseFull": "$collapseGallery",
-        "copiedToClipboard": "$copiedToClipboard",
-        "copyToClipboard": "$copyToClipboard",
-        "description": "$description",
-        "expand": "$expandTranscribedText",
-        "expandFull": "$expandGallery",
-        "holdingText": "$moduleGoesHere",
-        "less": "$less",
-        "license": "$license",
-        "logo": "$logo",
-        "manifestHeader": "$aboutTheItem",
-        "more": "$more",
-        "noData": "$noDataToDisplay",
-        "page": "$page",
-        "rangeHeader": "$aboutThisSection",
-        "title": "$transcribedText",
-        "leftPage": "$transcribedTextLeftPage",
-        "rightPage": "$transcribedTextRightPage",
-        "textNotFound": "$noTranscribedTextFound"
       }
     },
     "centerPanel": {
