+++ src/content-handlers/iiif/extensions/uv-openseadragon-extension/config/Config.ts
@@ -19,6 +19,10 @@ import {
 } from "@/content-handlers/iiif/BaseConfig";
 
 import { ContentLeftPanel } from "../../config/ContentLeftPanel";
+import {
+  ExpandPanelContent,
+  ExpandPanelOptions,
+} from "../../config/ExpandPanel";
 
 type MultiSelectDialogueOptions = DialogueOptions & {
   /** Determines if chunked resizing is enabled for gallery thumbnails */
@@ -163,6 +167,123 @@ type OpenSeadragonCenterPanelContent = CenterPanelContent & {
 type OpenSeadragonCenterPanel = ModuleConfig & {
   options: OpenSeadragonCenterPanelOptions;
   content: OpenSeadragonCenterPanelContent;
+};
+
+type SearchLeftPanelOptions = DialogueOptions &
+  ExpandPanelOptions & {
+    /** Order in which canvases are displayed */
+    canvasDisplayOrder: string;
+    /** Canvases to exclude from display */
+    canvasExclude: string;
+    /** Determines if copying to clipboard is enabled */
+    copyToClipboardEnabled: boolean;
+    /** Determines if download is enabled */
+    limitToRange: boolean;
+    /** Order in which manifests are displayed */
+    manifestDisplayOrder: string;
+    /** Manifests to exclude from display */
+    manifestExclude: string;
+    /** Language codes for right-to-left languages */
+    rtlLanguageCodes: string;
+    /** Determines if all languages should be shown */
+    showAllLanguages: boolean;
+    /** Limit for the text */
+    textLimit: number;
+    /** Type of the text limit */
+    textLimitType: string;
+  };
+
+type SearchLeftPanelContent = DialogueContent &
+  ExpandPanelContent & {
+    attribution: string;
+    canvasHeader: string;
+    collapse: string;
+    collapseFull: string;
+    copiedToClipboard: string;
+    copyToClipboard: string;
+    description: string;
+    expand: string;
+    expandFull: string;
+    holdingText: string;
+    less: string;
+    license: string;
+    logo: string;
+    manifestHeader: string;
+    more: string;
+    noData: string;
+    page: string;
+    rangeHeader: string;
+    title: string;
+    enterKeyword: string;
+    searchWithin: string;
+    clearSearch: string;
+    doSearch: string;
+    instanceFound: string;
+    instancesFound: string;
+    noMatches: string;
+    nextResult: string;
+    previousResult: string;
+    searchHitOf: string;
+  };
+
+export type SearchLeftPanel = ModuleConfig & {
+  options: SearchLeftPanelOptions;
+  content: SearchLeftPanelContent;
+};
+
+type TextRightPanelOptions = DialogueOptions &
+  ExpandPanelOptions & {
+    /** Order in which canvases are displayed */
+    canvasDisplayOrder: string;
+    /** Canvases to exclude from display */
+    canvasExclude: string;
+    /** Determines if copying to clipboard is enabled */
+    copyToClipboardEnabled: boolean;
+    /** Determines if download is enabled */
+    limitToRange: boolean;
+    /** Order in which manifests are displayed */
+    manifestDisplayOrder: string;
+    /** Manifests to exclude from display */
+    manifestExclude: string;
+    /** Language codes for right-to-left languages */
+    rtlLanguageCodes: string;
+    /** Determines if all languages should be shown */
+    showAllLanguages: boolean;
+    /** Limit for the text */
+    textLimit: number;
+    /** Type of the text limit */
+    textLimitType: string;
+  };
+
+type TextRightPanelContent = DialogueContent &
+  ExpandPanelContent & {
+    attribution: string;
+    canvasHeader: string;
+    collapse: string;
+    collapseFull: string;
+    copiedToClipboard: string;
+    copyToClipboard: string;
+    description: string;
+    expand: string;
+    expandFull: string;
+    holdingText: string;
+    less: string;
+    license: string;
+    logo: string;
+    manifestHeader: string;
+    more: string;
+    noData: string;
+    page: string;
+    rangeHeader: string;
+    title: string;
+    leftPage: string;
+    rightPage: string;
+    textNotFound: string;
+  };
+
+export type TextRightPanel = ModuleConfig & {
+  options: TextRightPanelOptions;
+  content: TextRightPanelContent;
 };
 
 type SearchFooterPanelOptions = FooterPanelOptions & {
@@ -315,6 +436,8 @@ type OSDSettingsDialogue = ModuleConfig & {
 
 type Modules = {
   contentLeftPanel: ContentLeftPanel;
+  searchLeftPanel: SearchLeftPanel;
+  textRightPanel: TextRightPanel;
   downloadDialogue: OSDDownloadDialogue;
   multiSelectDialogue: MultiSelectDialogue;
   pagingHeaderPanel: PagingHeaderPanel;
