+++ src/content-handlers/iiif/modules/uv-shared-module/css/sm.less
@@ -14,6 +14,16 @@
             visibility: visible;
         }
 
+        .searchLeftPanel {
+          display: none;
+          visibility: hidden;
+        }
+
+        .textRightPanel {
+          display: none;
+          visibility: hidden;
+        }
+
         .footerPanel {
             display: none;
             visibility: hidden;
