+++ src/locales/cy-GB.json
@@ -90,8 +90,6 @@
   "$lastImage": "Delwedd Olaf",
   "$lastPage": "Tudalen Olaf",
   "$next": "Nesaf",
-  "$nextFiveImages": "Forward five images",
-  "$nextFivePages": "Forward five pages",
   "$nextImage": "Delwedd Nesaf",
   "$nextPage": "Tudalen Nesaf",
   "$of": "o {0}",
@@ -163,15 +161,5 @@
   "$brightness": "Disgleirdeb",
   "$contrast": "Cyferbyniad",
   "$saturation": "Dirlawnder",
-  "$transcribedText": "Transcribed text",
-  "$expandTranscribedText": "Expand transcribed text",
-  "$collapseTranscribedText": "Collapse transcribed text",
-  "$transcribedTextLeftPage": "Left page",
-  "$transcribedTextRightPage": "Right page",
-  "$noTranscribedTextFound": "Couldn't find any transcriptions for this page.",
-  "$search": "Search",
-  "$expandSearch": "Expand Search",
-  "$collapseSearch": "Collapse Search",
-  "$searchHitOf": "{0} of {1}",
   "$imageSelectionLabel": "Dewiswch dudalen/delwedd"
 }
