+++ src/locales/fr-FR.json
@@ -98,8 +98,6 @@
   "$singlePageView": "Affichage en mode page simple",
   "$searchByPageNumber": "Recherche par numéro de page.",
   "$previous": "Précédent",
-  "$previousFiveImages": "Back five images",
-  "$previousFivePages": "Back five pages",
   "$previousImage": "Image précédente",
   "$previousPage": "Page précédent",
   "$settings": "Paramètres",
@@ -160,15 +158,5 @@
   "$mute": "Muet",
   "$pause": "Pause",
   "$play": "Jouer",
-  "$transcribedText": "Transcribed text",
-  "$expandTranscribedText": "Expand transcribed text",
-  "$collapseTranscribedText": "Collapse transcribed text",
-  "$transcribedTextLeftPage": "Left page",
-  "$transcribedTextRightPage": "Right page",
-  "$noTranscribedTextFound": "Couldn't find any transcriptions for this page.",
-  "$search": "Recherche",
-  "$expandSearch": "Expand Search",
-  "$collapseSearch": "Collapse Search",
-  "$searchHitOf": "{0} de {1}",
   "$imageSelectionLabel": "Choisissez une page/une image"
 }
