+++ src/locales/pl-PL.json
@@ -90,8 +90,6 @@
   "$lastImage": "Ostatni obraz",
   "$lastPage": "Ostatnia strona",
   "$next": "Następna",
-  "$nextFiveImages": "Forward five images",
-  "$nextFivePages": "Forward five pages",
   "$nextImage": "Następny obraz",
   "$nextPage": "Następna strona",
   "$of": "z {0}",
@@ -160,15 +158,5 @@
   "$mute": "Wycisz",
   "$pause": "Pauza",
   "$play": "Odtwarzaj",
-  "$transcribedText": "Transcribed text",
-  "$expandTranscribedText": "Expand transcribed text",
-  "$collapseTranscribedText": "Collapse transcribed text",
-  "$transcribedTextLeftPage": "Left page",
-  "$transcribedTextRightPage": "Right page",
-  "$noTranscribedTextFound": "Couldn't find any transcriptions for this page.",
-  "$search": "Search",
-  "$expandSearch": "Expand Search",
-  "$collapseSearch": "Collapse Search",
-  "$searchHitOf": "{0} z {1}",
   "$imageSelectionLabel": "Wybierz stronę/obraz"
 }
